+++ src/main/java/com/github/dakusui/floorplan/core/Connector.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.floorplan;
+package com.github.dakusui.floorplan.core;
 
 import com.github.dakusui.floorplan.component.Attribute;
 import com.github.dakusui.floorplan.component.Ref;
