+++ src/main/java/com/github/dakusui/floorplan/exception/Exceptions.java
@@ -1,6 +1,6 @@
 package com.github.dakusui.floorplan.exception;
 
-import com.github.dakusui.floorplan.FloorPlan;
+import com.github.dakusui.floorplan.core.FloorPlan;
 import com.github.dakusui.floorplan.component.Attribute;
 import com.github.dakusui.floorplan.component.Ref;
 import com.github.dakusui.floorplan.policy.Profile;
