+++ src/main/java/com/github/dakusui/floorplan/policy/Policy.java
@@ -1,8 +1,8 @@
 package com.github.dakusui.floorplan.policy;
 
-import com.github.dakusui.floorplan.Fixture;
-import com.github.dakusui.floorplan.FixtureConfigurator;
-import com.github.dakusui.floorplan.FloorPlan;
+import com.github.dakusui.floorplan.core.Fixture;
+import com.github.dakusui.floorplan.core.FixtureConfigurator;
+import com.github.dakusui.floorplan.core.FloorPlan;
 import com.github.dakusui.floorplan.component.Attribute;
 import com.github.dakusui.floorplan.component.ComponentSpec;
 import com.github.dakusui.floorplan.component.Configurator;
