+++ src/main/java/com/github/dakusui/floorplan/policy/Profile.java
@@ -1,6 +1,5 @@
 package com.github.dakusui.floorplan.policy;
 
-import com.github.dakusui.floorplan.Fixture;
 import com.github.dakusui.floorplan.component.Attribute;
 import com.github.dakusui.floorplan.component.Ref;
 import com.github.dakusui.floorplan.resolver.Resolver;
