+++ src/test/java/com/github/dakusui/floorplan/examples/bookstore/floorplan/BookstoreFixture.java
@@ -1,7 +1,7 @@
-package com.github.dakusui.floorplan.examples.bookstore;
+package com.github.dakusui.floorplan.examples.bookstore.floorplan;
 
-import com.github.dakusui.floorplan.Fixture;
-import com.github.dakusui.floorplan.FixtureConfigurator;
+import com.github.dakusui.floorplan.core.Fixture;
+import com.github.dakusui.floorplan.core.FixtureConfigurator;
 import com.github.dakusui.floorplan.policy.Policy;
 
 public abstract class BookstoreFixture extends Fixture.Base {
