+++ src/test/java/com/github/dakusui/floorplan/ut/tdesc/UtFixture.java
@@ -1,7 +1,7 @@
 package com.github.dakusui.floorplan.ut.tdesc;
 
-import com.github.dakusui.floorplan.Fixture;
-import com.github.dakusui.floorplan.FixtureConfigurator;
+import com.github.dakusui.floorplan.core.Fixture;
+import com.github.dakusui.floorplan.core.FixtureConfigurator;
 import com.github.dakusui.floorplan.policy.Policy;
 
 public class UtFixture extends Fixture.Base {
